requirements.txt
custom_utils==1.0.0
boto3==1.15.18
apache-airflow==2.10.3
snowflake-connector-python==3.12.4
apache-airflow-providers-snowflake==5.8.1
